pyproject.toml
[project]
name = "PixelateIt"
version = "0.1.0"
description = "A package to manipulate images"
authors = [
    {name = "Matthew D.",email = "mcdgit29@protonmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
tox = "^4.25.0"
pytest = "^8.3.5"
ruff = "^0.11.6"
flake8 = "^7.2.0"
sphinx = "^8.2.3"
black = "^25.1.0"
pre-commit = "^4.2.0"

[tool.poetry.extras]
dev = ["pytest"]